+++ CMP1005-A00133304-WebApi/MathOperationsWebApi/Controllers/MathOperationsController.cs
@@ -14,7 +14,7 @@ namespace MathOperationsWebApi.Controllers
     {
         [HttpGet]
         // Get the sum of 2 or more addends
-        public int Sum([FromQuery(Name = "addends[]")] int[] addends)
+        public decimal Sum([FromQuery(Name = "addends[]")] decimal[] addends)
         {
             return MathOperationsLib.MathOperationsLib.Sum(addends);
         }
